// backend/src/frontend-interfaces/beet-item.interface.ts
export interface BeetItemFE {
  id: number;
  year: number;
  track: number;
  tracktotal: number;
  disc: number;
  disctotal: number;
  bpm: number;
  original_year: number;
  bitrate: number;
  samplerate: number;
  bitdepth: number;
  channels: number;
  length: number;
  mtime: number;
  added: number;
  title: string;
  artist: string;
  albumartist: string;
  genre: string;
  initial_key: string;
  format: string;
  trackdisambig: string;
}